refactor(app): extract chat store hydration
